+++ public/sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = 'cok-static-v5'
+const CACHE_NAME = 'cok-static-v6'
 const APP_SHELL = [
   './',
   './index.html',